Implements part of ga engine

+++ sara-engine/SaraGeneticAlgorithmsPlugin/src/org/sara/sarageneticalgorithmsplugin/mutation/AbstractMutation.java
@@ -2,12 +2,12 @@ package org.sara.sarageneticalgorithmsplugin.mutation;
 
 import org.sara.interfaces.algorithms.ga.mutation.IMutation;
 import org.sara.interfaces.algorithms.ga.chromosome.IChromosome;
-import org.sara.interfaces.algorithms.ga.chromosome.IPopulation;
+import org.sara.interfaces.algorithms.ga.population.IPopulation;
 
 public abstract class AbstractMutation implements IMutation {
 
     @Override
-    public void mutate(IPopulation population, float rate) {
+    public void mutate(IPopulation population, double rate) {
         int qtdeMutations = (int) (population.size() * rate);
         for (int i = 0; i < qtdeMutations; i++) {
             IChromosome chromosome = population.getRandomChromosome();
