+++ sara-engine/SaraGeneticAlgorithmsPlugin/src/org/sara/sarageneticalgorithmsplugin/selection/BestSelection.java
@@ -2,15 +2,16 @@ package org.sara.sarageneticalgorithmsplugin.selection;
 
 import org.sara.interfaces.algorithms.ga.selection.ISelection;
 import org.sara.sarageneticalgorithmsplugin.defaultoptions.chromosome.DefaultPopulation;
-import org.sara.interfaces.algorithms.ga.chromosome.IPopulation;
+import org.sara.interfaces.algorithms.ga.population.IPopulation;
 
 public class BestSelection implements ISelection {
 
     @Override
-    public IPopulation select(IPopulation population, float rate) {
+    public IPopulation select(IPopulation population, double rate) {
         int countSelections = (int) (population.size() * rate);
         IPopulation newPopulation = new DefaultPopulation();
         population.sortByFitness();
+        
         for (int i = 0; i < countSelections; i++) {
             newPopulation.addChromosome(population.getChromosome(i));
         }
